refactor(parser, parser/boolean, types/ast, types/ast/boolean): move data structure to types/ast and keep parser/eval logic in parser
